refactor: getCards에서의 에러 처리
