+++ src/hooks/useCard.js
@@ -1,12 +1,18 @@
 import { useEffect, useState } from 'react';
+
 import { getCards } from '../service/card';
 
 const useCard = () => {
   const [cardList, setCardList] = useState([]);
 
-  useEffect(async () => {
-    setCardList(await getCards());
-  }, [cardList]);
+  useEffect(() => {
+    const loadCards = async () => {
+      const cards = await getCards();
+      setCardList(cards);
+    };
+
+    loadCards();
+  }, []);
 
   return { cardList };
 };
